chore: 🤖 move util tests to util test directory
