+++ tests/util/universal.test.js
@@ -1,19 +1,17 @@
-import construct from '../src/lib/util/construct'
+import universal from '../../src/lib/util/universal'
 
 test('constructs array', () => {
-  expect(construct(123, {}, 'test-property', (value) => value)).toStrictEqual([
-    ['test-property', ':'],
-    123,
-    ';',
-  ])
   expect(
-    construct('asdasd', {}, 'test-property', (value) => value)
-  ).toStrictEqual([['test-property', ':'], 'asdasd', ';'])
+    universal(123, {}, 'test-property:test(', ');', (value) => value)
+  ).toStrictEqual(['test-property:test(', 123, ');'])
+  expect(
+    universal('asdasd', {}, 'test-property:test(', ');', (value) => value)
+  ).toStrictEqual(['test-property:test(', 'asdasd', ');'])
 })
 
 test('constructs array with breakpoints', () => {
   expect(
-    construct(
+    universal(
       {
         standard: 12,
         custom: 22,
@@ -31,59 +29,61 @@ test('constructs array with breakpoints', () => {
           wide: '1440px',
         },
       },
-      'test-property',
+      'test-property:test(',
+      ');',
       (value) => value
     )
   ).toStrictEqual([
-    [['test-property', ':'], 12, ';'],
+    ['test-property:test(', 12, ');'],
     [
       ['@media screen and (min-width:', '200px', ')'],
       '{',
-      ['test-property', ':'],
+      'test-property:test(',
       22,
-      ';',
+      ');',
       '}',
     ],
     [
       ['@media screen and (min-width:', '320px', ')'],
       '{',
-      ['test-property', ':'],
+      'test-property:test(',
       '34px',
-      ';',
+      ');',
       '}',
     ],
     [
       ['@media screen and (min-width:', '768px', ')'],
       '{',
-      ['test-property', ':'],
+      'test-property:test(',
       '41',
-      ';',
+      ');',
       '}',
     ],
     [
       ['@media screen and (min-width:', '1024px', ')'],
       '{',
-      ['test-property', ':'],
+      'test-property:test(',
       55,
-      ';',
+      ');',
       '}',
     ],
     [
       ['@media screen and (min-width:', '1440px', ')'],
       '{',
-      ['test-property', ':'],
+      'test-property:test(',
       '67rem',
-      ';',
+      ');',
       '}',
     ],
   ])
   expect(
-    construct(
+    universal(
       ['22rem', 34, 41, '55px', '67ch'],
       {
         breakpoints: ['200px', '320px', '768px', '1024px', '1440px'],
       },
-      'test-property',
+      'test-property:test(',
+      ');',
       (value) => value
     )
   ).toStrictEqual([
@@ -91,41 +91,41 @@ test('constructs array with breakpoints', () => {
     [
       ['@media screen and (min-width:', '200px', ')'],
       '{',
-      ['test-property', ':'],
+      'test-property:test(',
       '22rem',
-      ';',
+      ');',
       '}',
     ],
     [
       ['@media screen and (min-width:', '320px', ')'],
       '{',
-      ['test-property', ':'],
+      'test-property:test(',
       34,
-      ';',
+      ');',
       '}',
     ],
     [
       ['@media screen and (min-width:', '768px', ')'],
       '{',
-      ['test-property', ':'],
+      'test-property:test(',
       41,
-      ';',
+      ');',
       '}',
     ],
     [
       ['@media screen and (min-width:', '1024px', ')'],
       '{',
-      ['test-property', ':'],
+      'test-property:test(',
       '55px',
-      ';',
+      ');',
       '}',
     ],
     [
       ['@media screen and (min-width:', '1440px', ')'],
       '{',
-      ['test-property', ':'],
+      'test-property:test(',
       '67ch',
-      ';',
+      ');',
       '}',
     ],
   ])
@@ -133,7 +133,7 @@ test('constructs array with breakpoints', () => {
 
 test('constructs array with mediaQueries', () => {
   expect(
-    construct(
+    universal(
       {
         standard: 11,
         custom: '25px',
@@ -151,19 +151,20 @@ test('constructs array with mediaQueries', () => {
           wide: '@test and wide',
         },
       },
-      'test-property',
+      'test-property:test(',
+      ');',
       (value) => value
     )
   ).toStrictEqual([
-    [['test-property', ':'], 11, ';'],
-    ['@test and custom', '{', ['test-property', ':'], '25px', ';', '}'],
-    ['@test and mobile', '{', ['test-property', ':'], 39, ';', '}'],
-    ['@test and tablet', '{', ['test-property', ':'], '44ch', ';', '}'],
-    ['@test and desktop', '{', ['test-property', ':'], '51rem', ';', '}'],
-    ['@test and wide', '{', ['test-property', ':'], 64, ';', '}'],
+    ['test-property:test(', 11, ');'],
+    ['@test and custom', '{', 'test-property:test(', '25px', ');', '}'],
+    ['@test and mobile', '{', 'test-property:test(', 39, ');', '}'],
+    ['@test and tablet', '{', 'test-property:test(', '44ch', ');', '}'],
+    ['@test and desktop', '{', 'test-property:test(', '51rem', ');', '}'],
+    ['@test and wide', '{', 'test-property:test(', 64, ');', '}'],
   ])
   expect(
-    construct(
+    universal(
       ['24px', 31, '43ch', 52, '62rem'],
       {
         mediaQueries: [
@@ -174,15 +175,16 @@ test('constructs array with mediaQueries', () => {
           '@test and 4',
         ],
       },
-      'test-property',
+      'test-property:test(',
+      ');',
       (value) => value
     )
   ).toStrictEqual([
     false,
-    ['@test and 0', '{', ['test-property', ':'], '24px', ';', '}'],
-    ['@test and 1', '{', ['test-property', ':'], 31, ';', '}'],
-    ['@test and 2', '{', ['test-property', ':'], '43ch', ';', '}'],
-    ['@test and 3', '{', ['test-property', ':'], 52, ';', '}'],
-    ['@test and 4', '{', ['test-property', ':'], '62rem', ';', '}'],
+    ['@test and 0', '{', 'test-property:test(', '24px', ');', '}'],
+    ['@test and 1', '{', 'test-property:test(', 31, ');', '}'],
+    ['@test and 2', '{', 'test-property:test(', '43ch', ');', '}'],
+    ['@test and 3', '{', 'test-property:test(', 52, ');', '}'],
+    ['@test and 4', '{', 'test-property:test(', '62rem', ');', '}'],
   ])
 })
